refactor(codegen): give callee identity one home and thin out ast.ts
Five helpers across five files answered "what is this callee called", with three different trust
models, no cross references, and nothing to steer a feeder to the right one — the drift that
already produced one live disagreement on the RLS signal earlier in this branch.

They collapse to three in `discover/callee.ts`, documented as a deliberate ladder from most to
least trusting: `resolveCalleeKind` believes only a surface-gated import (so a local
`const query = …` is not a registration), `resolvesToImportedName` also accepts anything spelled
right (so the middleware detectors keep working under degraded type info), and `calleeName`
answers the spelling alone and makes no claim about origin. Each docblock now says which of the
others it relaxes and what that costs.

Three of the five were the same function with cosmetic differences — a `""` fallback instead of
`undefined`, or taking the call rather than the callee — and are gone. The fifth, in
`privileged-dispatches.ts`, genuinely differs: it accepts a bare identifier only, because a
handler factory is never called as a member and accepting `x.defineQueue(...)` would let an
unrelated call register a privileged dispatch. It stays, renamed `bareCalleeName` so the
difference is visible at the call site rather than discovered later.

`ast.ts` drops from 22 exports to 17 and 522 lines to 385: the builder-chain walk moves to
`discover/builder-chain.ts` beside the other chain code, `resolvesToImportedName` was never a
generic AST primitive, and `isDatabaseCall` was exported despite having no caller outside the
file. The source-set walkers stay — `listLunoraSourceFiles` alone has 43 importers, and moving
them would be fifty files of churn for a better filename.
