+++ packages/codegen/src/discover/ai-tool-side-effects.ts
@@ -1,9 +1,10 @@
 import type { CallExpression, Node as TsNode, ObjectLiteralExpression, Project, SourceFile } from "ts-morph";
 import { Node, SyntaxKind } from "ts-morph";
 
-import { calleeName, enclosingExportName, referencesArgs } from "../argument-taint";
+import { enclosingExportName, referencesArgs } from "../argument-taint";
 import type { AiToolSideEffectIR } from "../ir";
 import { listLunoraSourceFiles, lunoraRelativePath } from "./ast";
+import { calleeName } from "./callee";
 
 /** The AI SDK text-generation entrypoints that accept a `tools` map — the injection sink surface. Matched by callee name, `import`-agnostic like the other feeders. */
 const GENERATION_CALLEES = new Set(["generateText", "streamText"]);
