+++ packages/codegen/src/discover/rls-procedures/internal/chain.ts
@@ -1,7 +1,8 @@
 import type { CallExpression, Node as TsNode } from "ts-morph";
 import { Node } from "ts-morph";
 
-import { resolvesToImportedName, wrappedCallsInChain } from "../../ast";
+import { wrappedCallsInChain } from "../../builder-chain";
+import { resolvesToImportedName } from "../../callee";
 
 /**
  * True when `node` is a `CallExpression` whose callee resolves to the name
